api/requirements.txt
sanic
mysql-connector-python